+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - bug fixes
     - custom module's submodules not loading their own submodules with relative paths properly
     - harmless error message when importing css files from the main theme file
+    - allow using folder names "client" and "assets" for user files (eg for images used in a session) 
 
 ## 1.16.2