+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## 1.16.3
 
 - bug fixes
-    - custom module's submodules not loading their own submodules with relative paths properly
-    - harmless error message when importing css files from the main theme file
-    - allow using folder names "client" and "assets" for user files (eg for images used in a session) 
+    - custom module: submodules not loading their own submodules with relative paths properly
+    - custom module: issue with circular submodule requires
+    - server: harmless error message when importing css files from the main theme file
+    - server: allow using folder names "client" and "assets" for user files (eg for images used in a session) 
 
 ## 1.16.2